Collapse the two URL resolvers into one shared core
resolveUrl in action/feeds/parsers.ts and resolveEntryUrl in lib/utils.ts
independently implemented the same rules -- trim, pass data: through, require
an http(s) base, new URL with a fallback -- and stayed in agreement only
because comments in each file said the other one did the same thing. They had
already drifted once: the scheme-less // rule was added to the action and the
reader was missed for a round.

Every difference between the two turned out to be a give-up case, and the two
disagreed only on what to hand back: the action the trimmed URL, the reader the
original. So the shared core returns null for "leave this alone" and each side
spells its own answer in a `??` at the call site. The // rule takes no base and
stays out of the core entirely, which puts both deliberate deviations -- the
action inheriting the feed's scheme, the reader forcing https -- where a reader
of either function can see them.

Renames the confusable pair. absoluteEntryLink makes the entry's own link
absolute; resolveEntryUrl resolves some other URL against the entry. Both read
as "the entry's URL" and did close to opposite things, so they become
absolutizeEntryLink and resolveAgainstEntry.

lib/media.ts becomes lib/entry-urls.ts now that it owns link resolution as well
as media, and stops colliding with action/feeds/media.ts, which downloads files.
The reader's resolver moves there out of lib/utils.ts, a module of page-state
class-name helpers that also imports ./storage -- the action's parser tests were
pulling the reader's storage layer in to check one string function.

Two edges change, both previously unpinned, and both remove a divergence:
an absolute URL in content with no usable base is no longer re-serialized by
the action, and //x/y with no entry URL is now resolved by the reader.
