+++ action/feeds/parsers.ts
@@ -1,7 +1,12 @@
 import { parseString } from 'xml2js'
 import sanitizeHtml from 'sanitize-html'
 
-import { mapUrlAttributes, type UrlTarget } from '../../lib/media'
+import {
+  mapUrlAttributes,
+  parseHttpUrl,
+  resolveAgainstBase,
+  type UrlTarget
+} from '../../lib/entry-urls'
 
 export interface Entry {
   title: string
@@ -29,47 +34,37 @@ function joinValuesOrEmptyString(values: Values) {
   return (values && values.join('').trim()) || ''
 }
 
-function parseAbsoluteHttpUrl(input?: string | null) {
-  if (!input) return null
-  try {
-    const parsed = new URL(input)
-    if (!['http:', 'https:'].includes(parsed.protocol)) return null
-    return parsed
-  } catch {
-    return null
-  }
-}
-
+/**
+ * Resolves a URL against the two bases a feed offers, preferring the first and
+ * falling back to the second. A URL that resolves against neither is handed
+ * back trimmed, which is where the action parts from the reader: the reader
+ * returns the input untouched instead, since it has no sanitizer downstream to
+ * drop a blank attribute for it.
+ */
 function resolveUrl(
   inputUrl: string,
   primaryBase: string,
   fallbackBase: string
 ) {
   const trimmed = inputUrl.trim()
-  if (!trimmed) return trimmed
-  if (trimmed.startsWith('data:')) return trimmed
 
+  // Deliberately not the reader's rule, which pins these to https. A
+  // scheme-less URL takes no base, so this is decided here rather than in
+  // resolveAgainstBase: the action inherits the feed's own scheme, and only
+  // falls back to https when the feed offers no usable base to inherit from.
   if (trimmed.startsWith('//')) {
     const protocol =
-      parseAbsoluteHttpUrl(primaryBase)?.protocol ||
-      parseAbsoluteHttpUrl(fallbackBase)?.protocol ||
+      parseHttpUrl(primaryBase)?.protocol ||
+      parseHttpUrl(fallbackBase)?.protocol ||
       'https:'
     return `${protocol}${trimmed}`
   }
 
-  const absolute = parseAbsoluteHttpUrl(trimmed)
-  if (absolute) return absolute.toString()
-
-  const base =
-    parseAbsoluteHttpUrl(primaryBase)?.toString() ||
-    parseAbsoluteHttpUrl(fallbackBase)?.toString()
-  if (!base) return trimmed
-
-  try {
-    return new URL(trimmed, base).toString()
-  } catch {
-    return trimmed
-  }
+  return (
+    resolveAgainstBase(trimmed, primaryBase) ??
+    resolveAgainstBase(trimmed, fallbackBase) ??
+    trimmed
+  )
 }
 
 /**
@@ -84,8 +79,8 @@ function resolveUrl(
  * run after this ships: its stored entries reappear under new keys and the old
  * ones are cleaned up. That is the cost of the fix, not a bug to undo.
  */
-function absoluteEntryLink(rawLink: string, siteLink: string) {
-  if (!rawLink || parseAbsoluteHttpUrl(rawLink)) return rawLink
+function absolutizeEntryLink(rawLink: string, siteLink: string) {
+  if (!rawLink || parseHttpUrl(rawLink)) return rawLink
   return resolveUrl(rawLink, siteLink, '')
 }
 
@@ -136,9 +131,10 @@ export const ENTRY_CONTENT_SANITIZE_OPTIONS: sanitizeHtml.IOptions = {
   allowedTags: sanitizeHtml.defaults.allowedTags.concat(['img']),
   allowedAttributes: {
     // An attribute added here that carries a URL has to be listed in
-    // URL_ATTRIBUTES in lib/media.ts too, or it keeps whatever relative URL the
-    // feed published and lands on the reader's own domain -- and its tag needs
-    // an allowedSchemesByTag entry below if http(s) is not the right set.
+    // URL_ATTRIBUTES in lib/entry-urls.ts too, or it keeps whatever relative
+    // URL the feed published and lands on the reader's own domain -- and its
+    // tag needs an allowedSchemesByTag entry below if http(s) is not the right
+    // set.
     ...sanitizeHtml.defaults.allowedAttributes,
     img: ['src', 'alt', 'title', 'width', 'height', 'loading', 'srcset'],
     // Kept so the source of a quotation survives into the stored JSON, and
@@ -246,7 +242,7 @@ export function parseRss(feedTitle: string, xml: any): Site {
             pubDate,
             description: entryDescription
           } = item
-          const entryLink = absoluteEntryLink(
+          const entryLink = absolutizeEntryLink(
             joinValuesOrEmptyString(entryLinks),
             siteLink
           )
@@ -295,7 +291,7 @@ export function parseAtom(feedTitle: string, xml: any): Site {
             : summary
               ? summary[0]._
               : ''
-          const entryLink = absoluteEntryLink(
+          const entryLink = absolutizeEntryLink(
             (itemLink && itemLink.$.href) || '',
             siteUrl
           )
