+++ lib/utils.test.ts
@@ -1,89 +1,5 @@
 import test from 'ava'
-import { parseLocation, resolveEntryUrl } from './utils'
-
-const ENTRY_URL = 'https://feed.example/posts/entry-1'
-
-test('#resolveEntryUrl resolves relative urls against the entry', (t) => {
-  t.is(
-    resolveEntryUrl('/posts/other', ENTRY_URL),
-    'https://feed.example/posts/other'
-  )
-  t.is(
-    resolveEntryUrl('chapter-two.html', ENTRY_URL),
-    'https://feed.example/posts/chapter-two.html'
-  )
-  t.is(
-    resolveEntryUrl('#footnote', ENTRY_URL),
-    'https://feed.example/posts/entry-1#footnote'
-  )
-  t.is(
-    resolveEntryUrl('//en.wikipedia.org/wiki/RSS', ENTRY_URL),
-    'https://en.wikipedia.org/wiki/RSS'
-  )
-  // A scheme-less URL takes the scheme of the page it ends up on, which is the
-  // reader, so an http entry does not drag it down to plaintext -- the same
-  // rule the action applies when it resolves content.
-  t.is(
-    resolveEntryUrl('//x.example/y', 'http://feed.example/p'),
-    'https://x.example/y'
-  )
-})
-
-test('#resolveEntryUrl keeps urls it must not rewrite', (t) => {
-  t.is(
-    resolveEntryUrl('https://other.example/page', ENTRY_URL),
-    'https://other.example/page'
-  )
-  t.is(
-    resolveEntryUrl('mailto:user@example.com', ENTRY_URL),
-    'mailto:user@example.com'
-  )
-  // A payload the URL parser would rewrite, so the assertion can tell being
-  // handed back untouched apart from being round-tripped through it.
-  t.is(
-    resolveEntryUrl('data:image/svg+xml,<svg>café</svg>', ENTRY_URL),
-    'data:image/svg+xml,<svg>café</svg>'
-  )
-  // Signed CDN URLs carry characters a stricter normalizer would escape.
-  t.is(
-    resolveEntryUrl('https://cdn.example/x?sig=a|b^c', ENTRY_URL),
-    'https://cdn.example/x?sig=a|b^c'
-  )
-})
-
-test('#resolveEntryUrl re-serializes an absolute url', (t) => {
-  // Every URL in stored content goes through here now, so the normalizing is
-  // worth pinning: it keeps the target but not necessarily the exact bytes.
-  t.is(
-    resolveEntryUrl('https://other.example', ENTRY_URL),
-    'https://other.example/'
-  )
-  t.is(
-    resolveEntryUrl('HTTPS://Other.Example/Page', ENTRY_URL),
-    'https://other.example/Page'
-  )
-})
-
-test('#resolveEntryUrl returns the url unchanged without a usable entry', (t) => {
-  t.is(resolveEntryUrl('/posts/other'), '/posts/other')
-  t.is(resolveEntryUrl('/posts/other', ''), '/posts/other')
-  t.is(resolveEntryUrl('/posts/other', 'not a url'), '/posts/other')
-  t.is(resolveEntryUrl('', ENTRY_URL), '')
-  t.is(resolveEntryUrl('   ', ENTRY_URL), '   ')
-})
-
-test('#resolveEntryUrl refuses a base that is not http', (t) => {
-  // The entry link is whatever the feed published. Resolving against a
-  // javascript: base would turn every footnote anchor into a script URL.
-  t.is(resolveEntryUrl('#fn1', 'javascript:alert(1)'), '#fn1')
-  t.is(resolveEntryUrl('#fn1', 'file:///etc/passwd'), '#fn1')
-  t.is(resolveEntryUrl('#fn1', 'data:text/html,<script></script>'), '#fn1')
-  t.is(resolveEntryUrl('/posts/other', 'ftp://feed.example/x'), '/posts/other')
-  t.is(
-    resolveEntryUrl('/posts/other', 'http://feed.example/posts/1'),
-    'http://feed.example/posts/other'
-  )
-})
+import { parseLocation } from './utils'
 
 test('#parseLocation returns category type', (t) => {
   t.deepEqual(parseLocation('/categories/Apple'), {
